move computed to marketStore
